test(cli): share bootstrap test helpers
